refactor(pages): organize page files into feature-based folders
