+++ event-planner/frontend/src/pages/events/EventDetails.css
