+++ event-planner/frontend/src/pages/home/Home.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import EventList from '../components/event/EventList';
+import EventList from '../../components/event/EventList';
 import './Home.css';
 
 const Home: React.FC = () => {
